FROMGIT: net: ipa: kill IPA_MEM_UC_OFFSET
The microcontroller shared memory area is at the beginning of the
IPA resident memory.  IPA_MEM_UC_OFFSET was defined as the offset
within that region where it's found, but it's 0, and it's never
actually used.  Just get rid of the definition, and move some of the
description it had to be above the definition of the ipa_uc_mem_area
structure.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 722208ea3e2a19f927cd507366547b2546444697
 git://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git master)

BUG=b:74020136
TEST=Booted trogdor, modem call functional

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: I14bd93d38a5b2d7555c82011408b461eee313109
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2283978
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>