FROMGIT: nvmem: core: add support to auto devid
For nvmem providers which have multiple instances, it is required
to suffix the provider name with proper id, so that they do not
confict for the same name. Currently the core does not handle
this case properly eventhough core already has logic to generate the id.

This patch add new devid type NVMEM_DEVID_AUTO for providers to be
able to allow core to assign id and append it to provier name.

Reported-by: Shawn Guo <shawn.guo@linaro.org>
Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
Tested-by: Shawn Guo <shawn.guo@linaro.org>
(cherry picked from commit 93ac5fdba1eddc679e9694b64f2fa321317df988
 git://git.kernel.org/pub/scm/linux/kernel/git/srini/nvmem.git for-next)

BUG=b:161344895
TEST=Later patch compiles

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I57317c1f0184b1a31da9a10e7e29ca98eddad18f
Disallow-Recycled-Builds: kukui-arc-r-cq
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2309058
Reviewed-by: Sean Paul <seanpaul@chromium.org>